backend/requirements.txt
# Task ID: T-016
fastapi
sqlmodel
uvicorn[standard]
openai
mcp
python-dotenv
psycopg2-binary
requests